Seperate UVM.go into Host/Container/Process
