+++ internal/runtime/hcsv2/uvm.go
@@ -5,13 +5,10 @@ package hcsv2
 import (
 	"bufio"
 	"encoding/json"
-	"fmt"
 	"io/ioutil"
 	"os"
-	"os/exec"
 	"path"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"sync"
 	"syscall"
@@ -25,7 +22,6 @@ import (
 	"github.com/Microsoft/opengcs/service/gcs/gcserr"
 	"github.com/Microsoft/opengcs/service/gcs/prot"
 	"github.com/Microsoft/opengcs/service/gcs/runtime"
-	"github.com/Microsoft/opengcs/service/gcs/stdio"
 	"github.com/Microsoft/opengcs/service/gcs/transport"
 	"github.com/opencontainers/runc/libcontainer/devices"
 	oci "github.com/opencontainers/runtime-spec/specs-go"
@@ -416,363 +412,4 @@ func (h *Host) ModifyHostSettings(settings *prot.ModifySettingRequest) error {
 // state that has not been cleaned before calling this function.
 func (h *Host) Shutdown() {
 	syscall.Reboot(syscall.LINUX_REBOOT_CMD_POWER_OFF)
-}
-
-type Container struct {
-	id    string
-	vsock transport.Transport
-
-	spec *oci.Spec
-
-	container   runtime.Container
-	initProcess *Process
-
-	processesMutex sync.Mutex
-	processesWg    sync.WaitGroup
-	processes      map[uint32]*Process
-}
-
-func (c *Container) Start(conSettings stdio.ConnectionSettings) (int, error) {
-	logrus.WithFields(logrus.Fields{
-		"cid": c.id,
-	}).Info("opengcs::Container::Start")
-
-	stdioSet, err := stdio.Connect(c.vsock, conSettings)
-	if err != nil {
-		return -1, err
-	}
-	if c.initProcess.spec.Terminal {
-		ttyr := c.container.Tty()
-		ttyr.ReplaceConnectionSet(stdioSet)
-		ttyr.Start()
-	} else {
-		pr := c.container.PipeRelay()
-		pr.ReplaceConnectionSet(stdioSet)
-		pr.CloseUnusedPipes()
-		pr.Start()
-	}
-	err = c.container.Start()
-	if err != nil {
-		stdioSet.Close()
-	}
-	return int(c.initProcess.pid), err
-}
-
-func (c *Container) ExecProcess(process *oci.Process, conSettings stdio.ConnectionSettings) (int, error) {
-	logrus.WithFields(logrus.Fields{
-		"cid": c.id,
-	}).Info("opengcs::Container::ExecProcess")
-
-	stdioSet, err := stdio.Connect(c.vsock, conSettings)
-	if err != nil {
-		return -1, err
-	}
-
-	// Increment the waiters before the call so that WaitContainer cannot complete in a race
-	// with adding a new process. When the process exits it will decrement this count.
-	c.processesMutex.Lock()
-	c.processesWg.Add(1)
-	c.processesMutex.Unlock()
-
-	p, err := c.container.ExecProcess(process, stdioSet)
-	if err != nil {
-		// We failed to exec any process. Remove our early count increment.
-		c.processesMutex.Lock()
-		c.processesWg.Done()
-		c.processesMutex.Unlock()
-		stdioSet.Close()
-		return -1, err
-	}
-
-	pid := p.Pid()
-	c.processesMutex.Lock()
-	c.processes[uint32(pid)] = newProcess(c, process, p, uint32(pid))
-	c.processesMutex.Unlock()
-	return pid, nil
-}
-
-// GetProcess returns the *Process with the matching 'pid'. If the 'pid' does
-// not exit returns error.
-func (c *Container) GetProcess(pid uint32) (*Process, error) {
-	logrus.WithFields(logrus.Fields{
-		"cid": c.id,
-		"pid": pid,
-	}).Info("opengcs::Container::GetProcess")
-
-	if c.initProcess.pid == pid {
-		return c.initProcess, nil
-	}
-
-	c.processesMutex.Lock()
-	defer c.processesMutex.Unlock()
-
-	p, ok := c.processes[pid]
-	if !ok {
-		return nil, errors.WithStack(gcserr.NewProcessDoesNotExistError(int(pid)))
-	}
-	return p, nil
-}
-
-// GetAllProcessPids returns all process pids in the container namespace.
-func (c *Container) GetAllProcessPids() ([]int, error) {
-	state, err := c.container.GetAllProcesses()
-	if err != nil {
-		return nil, err
-	}
-	pids := make([]int, len(state))
-	for i, s := range state {
-		pids[i] = s.Pid
-	}
-	return pids, nil
-}
-
-// Kill sends 'signal' to the container process.
-func (c *Container) Kill(signal syscall.Signal) error {
-	logrus.WithFields(logrus.Fields{
-		"cid":    c.id,
-		"signal": signal,
-	}).Info("opengcs::Container::Kill")
-
-	return c.container.Kill(signal)
-}
-
-// Wait waits for all processes exec'ed to finish as well as the init process
-// representing the container.
-func (c *Container) Wait() int {
-	logrus.WithFields(logrus.Fields{
-		"cid": c.id,
-	}).Info("opengcs::Container::Wait")
-
-	c.processesWg.Wait()
-	return c.initProcess.exitCode
-}
-
-// AddNetworkAdapter adds `a` to the network namespace held by this container.
-func (c *Container) AddNetworkAdapter(a *prot.NetworkAdapterV2) error {
-	log := logrus.WithFields(logrus.Fields{
-		"cid":               c.id,
-		"adapterInstanceID": a.ID,
-	})
-	log.Info("opengcs::Container::AddNetworkAdapter")
-
-	// TODO: netnscfg is not coded for v2 but since they are almost the same
-	// just convert the parts of the adapter here.
-	v1Adapter := &prot.NetworkAdapter{
-		NatEnabled:         a.IPAddress != "",
-		AllocatedIPAddress: a.IPAddress,
-		HostIPAddress:      a.GatewayAddress,
-		HostIPPrefixLength: a.PrefixLength,
-		EnableLowMetric:    a.EnableLowMetric,
-		EncapOverhead:      a.EncapOverhead,
-	}
-
-	cfg, err := json.Marshal(v1Adapter)
-	if err != nil {
-		return errors.Wrap(err, "failed to marshal adapter struct to JSON")
-	}
-
-	ifname, err := network.InstanceIDToName(a.ID, true)
-	if err != nil {
-		return err
-	}
-
-	out, err := exec.Command("netnscfg",
-		"-if", ifname,
-		"-nspid", strconv.Itoa(c.container.Pid()),
-		"-cfg", string(cfg)).CombinedOutput()
-	if err != nil {
-		return errors.Wrapf(err, "failed to configure adapter cid: %s, aid: %s, if id: %s %s", c.id, a.ID, ifname, string(out))
-	}
-	return nil
-}
-
-// RemoveNetworkAdapter removes the network adapter `id` from the network
-// namespace held by this container.
-func (c *Container) RemoveNetworkAdapter(id string) error {
-	logrus.WithFields(logrus.Fields{
-		"cid":               c.id,
-		"adapterInstanceID": id,
-	}).Info("opengcs::Container::RemoveNetworkAdapter")
-
-	// TODO: JTERRY75 - Implement removal if we ever need to support hot remove.
-	logrus.WithFields(logrus.Fields{
-		"cid":               c.id,
-		"adapterInstanceID": id,
-	}).Warning("opengcs::Container::RemoveNetworkAdapter - Not implemented")
-	return nil
-}
-
-// Process is a struct that defines the lifetime and operations associated with
-// an oci.Process.
-type Process struct {
-	spec *oci.Process
-	// cid is the container id that owns this process.
-	cid string
-
-	process runtime.Process
-	pid     uint32
-	// This is only valid post the exitWg
-	exitCode int
-	exitWg   sync.WaitGroup
-
-	// Used to allow addtion/removal to the writersWg after an initial wait has
-	// already been issued. It is not safe to call Add/Done without holding this
-	// lock.
-	writersSyncRoot sync.Mutex
-	// Used to track the number of writers that need to finish
-	// before the process can be marked for cleanup.
-	writersWg sync.WaitGroup
-	// Used to track the 1st caller to the writersWg that successfully
-	// acknowledges it wrote the exit response.
-	writersCalled bool
-}
-
-// newProcess returns a Process struct that has been initialized with an
-// outstanding wait for process exit, and post exit an outstanding wait for
-// process cleanup to release all resources once at least 1 waiter has
-// successfully written the exit response.
-func newProcess(c *Container, spec *oci.Process, process runtime.Process, pid uint32) *Process {
-	p := &Process{
-		spec:    spec,
-		process: process,
-		cid:     c.id,
-		pid:     pid,
-	}
-	p.exitWg.Add(1)
-	p.writersWg.Add(1)
-	go func() {
-		// Wait for the process to exit
-		exitCode, err := p.process.Wait()
-		if err != nil {
-			logrus.WithFields(logrus.Fields{
-				"cid":           c.id,
-				"pid":           pid,
-				logrus.ErrorKey: err,
-			}).Error("opengcs::Process - failed to wait for runc process")
-		}
-		p.exitCode = exitCode
-		logrus.WithFields(logrus.Fields{
-			"cid":      c.id,
-			"pid":      pid,
-			"exitCode": p.exitCode,
-		}).Info("opengcs::Process - process exited")
-
-		// Free any process waiters
-		p.exitWg.Done()
-		// Decrement any container process count waiters
-		c.processesMutex.Lock()
-		c.processesWg.Done()
-		c.processesMutex.Unlock()
-
-		// Schedule the removal of this process object from the map once at
-		// least one waiter has read the result
-		go func() {
-			p.writersWg.Wait()
-			c.processesMutex.Lock()
-
-			logrus.WithFields(logrus.Fields{
-				"cid": c.id,
-				"pid": pid,
-			}).Debug("opengcs::Process - all waiters have completed, removing process")
-
-			delete(c.processes, p.pid)
-			c.processesMutex.Unlock()
-		}()
-	}()
-	return p
-}
-
-// Kill sends 'signal' to the process.
-func (p *Process) Kill(signal syscall.Signal) error {
-	logrus.WithFields(logrus.Fields{
-		"cid":    p.cid,
-		"pid":    p.pid,
-		"signal": signal,
-	}).Info("opengcs::Process::Kill")
-
-	if err := syscall.Kill(int(p.pid), signal); err != nil {
-		return errors.WithStack(err)
-	}
-	return nil
-}
-
-// ResizeConsole resizes the tty to `height`x`width` for the process.
-func (p *Process) ResizeConsole(height, width uint16) error {
-	logrus.WithFields(logrus.Fields{
-		"cid":    p.cid,
-		"pid":    p.pid,
-		"height": height,
-		"width":  width,
-	}).Info("opengcs::Process::ResizeConsole")
-
-	tty := p.process.Tty()
-	if tty == nil {
-		return fmt.Errorf("pid: %d, is not a tty and cannot be resized", p.pid)
-	}
-	return tty.ResizeConsole(height, width)
-}
-
-// Wait returns a channel that can be used to wait for the process to exit and
-// gather the exit code. The second channel must be signaled from the caller
-// when the caller has completed its use of this call to Wait.
-func (p *Process) Wait() (<-chan int, chan<- bool) {
-	log := logrus.WithFields(logrus.Fields{
-		"cid": p.cid,
-		"pid": p.pid,
-	})
-	log.Info("opengcs::Process::Wait")
-
-	exitCodeChan := make(chan int, 1)
-	doneChan := make(chan bool)
-
-	// Increment our waiters for this waiter
-	p.writersSyncRoot.Lock()
-	p.writersWg.Add(1)
-	p.writersSyncRoot.Unlock()
-
-	go func() {
-		bgExitCodeChan := make(chan int, 1)
-		go func() {
-			p.exitWg.Wait()
-			bgExitCodeChan <- p.exitCode
-		}()
-
-		// Wait for the exit code or the caller to stop waiting.
-		select {
-		case exitCode := <-bgExitCodeChan:
-			exitCodeChan <- exitCode
-
-			// The caller got the exit code. Wait for them to tell us they have
-			// issued the write
-			select {
-			case <-doneChan:
-				p.writersSyncRoot.Lock()
-				// Decrement this waiter
-				log.Debug("opengcs::Process::Wait - wait completed, releasing wait count")
-
-				p.writersWg.Done()
-				if !p.writersCalled {
-					// We have at least 1 response for the exit code for this
-					// process. Decrement the release waiter that will free the
-					// process resources when the writersWg hits 0
-					log.Debug("opengcs::Process::Wait - first wait completed, releasing first wait count")
-
-					p.writersCalled = true
-					p.writersWg.Done()
-				}
-				p.writersSyncRoot.Unlock()
-			}
-
-		case <-doneChan:
-			// In this case the caller timed out before the process exited. Just
-			// decrement the waiter but since no exit code we just deal with our
-			// waiter.
-			p.writersSyncRoot.Lock()
-			log.Debug("opengcs::Process::Wait - wait canceled before exit, releasing wait count")
-
-			p.writersWg.Done()
-			p.writersSyncRoot.Unlock()
-		}
-	}()
-	return exitCodeChan, doneChan
 }
